[ALPS07087560] DISP: fix mmpprofile toool bug
MMP.exe support 64bit changelist patch:5182492 & 5182504

MTK-Commit-Id: 93ddc9c2c5abaab78c34454a01a8340e9836f982

CR-Id: ALPS07087560
Feature: mmp

+++ drivers/misc/mediatek/mmp/src/mmprofile.c
@@ -2239,8 +2239,8 @@ static long mmprofile_ioctl_compat(struct file *file, unsigned int cmd,
 
 static int mmprofile_mmap(struct file *file, struct vm_area_struct *vma)
 {
-	unsigned int pos = 0;
-	unsigned int i = 0;
+	unsigned long pos = 0;
+	unsigned long i = 0;
 
 	if (mmprofile_globals.selected_buffer == MMPROFILE_GLOBALS_BUFFER) {
